multiple addition due localization - adds translator to generated added properties of string type - simplifies translator - refactoring localization helper to AXSharp.Connector.Localization namespace
