+++ src/AXSharp.connectors/tests/AXSharp.ConnectorTests/AXSharp.ConnectorTests/Localizations/TranslatorExtensionTests.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
 using AXSharp.Connector.Localizations;
-using AXSharp.Localizations;
+
 
 
 namespace AXSharp.ConnectorTests.Localizations
